requirements.txt
openai>=1.25.0          # OpenRouter is OpenAI-compatible
httpx>=0.27.0
praw>=7.7.1             # Reddit API wrapper
apify-client>=1.7.0
python-dotenv>=1.0.0
rich>=13.7.0            # Pretty console output / logging
pydantic>=2.7.0         # Structured data models
jinja2>=3.1.4           # Report templating
tenacity>=8.3.0         # Retry logic for API calls
